requirements.txt
google-adk
google-genai
pydantic
python-dotenv
uvicorn
fastapi